Rename utility package to shared
